fix: RoomInfo 타입, InitialRoomInfo 값 별도 파일로 분리
